Moved section review heading to `ReviewHeading`; added edit button for review section - handler wip; TODOs added
